[C8] Split sketch.go by responsibility (#2229)
model/exchange/translators/inventor/ipt/sketch.go was 563 LOC, over the
500-line ceiling. Split the .ipt sketch reader along its responsibilities;
each file is now under the ceiling:

- sketch.go (141): the public sketch model (Point2D, Sketch, Line, Circle,
  Arc, Ellipse, Spline + IsConstruction accessors) and the decode entry
  points (DecodeSketch/DecodeSketches/clusterItems).
- sketch_records.go (248): the shared record decoding — forEachNamelessNode,
  collectItems (the byte-level entity scan), isEllipseNode, the curve-header
  consts, and the raw record types (idPoint/lineRefs/circleEnt/arcEnt/
  ellipseEnt/sketchItem).
- sketch_assemble.go (204): assembly + reference resolution — assembleCluster,
  resolveByRefs (the exact rank-align path) and the convex-ring fallback
  (assembleSketch/orderConvex/closeLoop).

Behaviour-preserving: every const, type and function body is byte-identical,
only moved; imports narrowed per file. The collectItems scan is a single pass
kept whole. The ipt corpus passes before and after (GOWORK=off, separate
module) with no golden updates; module lint issue count unchanged (32, the
pre-existing funlen/gocyclo/captLocal on collectItems/resolveByRefs travel
with their byte-identical functions). Net delta of recognizers / tolerance
constants / fallback sites / type assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2229.

(cherry picked from commit 064344674963073ecd0c2bccb7637bbc6763b185)
